renamed helpers to modals
